Consolidate package.from_uri tests
