+++ tests/ethpm/test_package_init.py
@@ -9,6 +9,7 @@ from ethpm import (
     Package,
 )
 from ethpm.exceptions import (
+    CannotHandleURI,
     EthPMValidationError,
 )
 
@@ -95,3 +96,36 @@ def test_from_file_succeeds_with_valid_manifest(valid_manifest_from_path, w3):
 def test_from_file_raises_type_error_with_invalid_param_type():
     with pytest.raises(TypeError):
         Package.from_file(1)
+
+
+#
+# From URI
+#
+
+VALID_IPFS_PKG = "ipfs://QmeD2s7KaBUoGYTP1eutHBmBkMMMoycdfiyGMx2DKrWXyV"
+
+
+def test_package_from_uri_with_valid_uri(dummy_ipfs_backend, w3):
+    package = Package.from_uri(VALID_IPFS_PKG, w3)
+    assert package.name == "safe-math-lib"
+    assert isinstance(package, Package)
+
+
+@pytest.mark.parametrize(
+    "uri",
+    (
+        # Invalid
+        "123",
+        b"123",
+        "ipfs://",
+        "http://QmTKB75Y73zhNbD3Y73xeXGjYrZHmaXXNxoZqGCagu7r8u/readme",
+        "ipfsQmTKB75Y73zhNbD3Y73xeXGjYrZHmaXXNxoZqGCagu7r8u/readme/",
+        # Unsupported
+        "erc111://packages.zeppelin.os/owned",
+        "bzz://da6adeeb4589d8652bbe5679aae6b6409ec85a20e92a8823c7c99e25dba9493d",
+    ),
+)
+@pytest.mark.skipif('WEB3_INFURA_PROJECT_ID' not in os.environ, reason='Infura API key unavailable')
+def test_package_from_uri_rejects_invalid_ipfs_uri(uri, w3):
+    with pytest.raises(CannotHandleURI):
+        Package.from_uri(uri, w3)
